requirements.txt
# Core ML Libraries
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
bitsandbytes>=0.41.0

# Web Interface
gradio>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0

# Embeddings (for semantic similarity in routing)
sentence-transformers>=2.5.0
faiss-cpu>=1.8.0

# LangGraph Multi-Agent Framework (Open-Source)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0  # Open-source local LLMs

# Optional: OpenAI/Groq (if API keys available)
langchain-openai>=0.2.0
langchain-groq>=0.2.0

# Open-Source Embeddings
sentence-transformers>=2.5.0

# Medical NER (optional BioBERT fallback)
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0

# Knowledge Graph
neo4j>=5.0.0
py2neo>=2021.2.3

# Additional utilities
beautifulsoup4>=4.12.0
loguru>=0.7.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

# Production API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
prometheus-client>=0.19.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0


